room creation and fetching works properly
